fix: redirect to splash or home
