+++ components/TopNav.vue
@@ -61,7 +61,7 @@ export default class TopNav extends Vue {
   async SignOut() {
     try {
       await authStore.signOut()
-      this.$router.push('/')
+      this.$router.push('/auth/Login')
     } catch (error) {}
   }
 }
